image/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
mangum==0.17.0
python-dotenv==1.0.0
openai>=1.40.0
python-multipart==0.0.6
ffmpeg-python==0.2.0
boto3==1.34.0
bcrypt==4.1.2
python-jose[cryptography]==3.3.0
